# Optimize stepper increment