Using @wolfv 's benchmark, "fixes" #695 

Without this change:

```
--------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations
--------------------------------------------------------------------------
dynamic_iterator<float>                  5006 ns       5004 ns     100000
iterator<float>                           805 ns        804 ns     798540
loop<float>                               943 ns        942 ns     646645
loop_view<float>                          720 ns        719 ns     969972
loop_raw<float>                           725 ns        724 ns     975515
assign<float>                            4664 ns       4662 ns     157197
assign_view<float>                       8336 ns       8334 ns      76757
assign_dynamic_view<float>              10955 ns      10951 ns      64198
assign_view_noalias<float>               4594 ns       4594 ns     149497
assign_dynamic_view_noalias<float>       6161 ns       6159 ns     109840
```

With this change
```
--------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations
--------------------------------------------------------------------------
dynamic_iterator<float>                  3089 ns       3085 ns     224263
iterator<float>                           709 ns        708 ns     911660
loop<float>                               929 ns        928 ns     688922
loop_view<float>                          686 ns        685 ns    1008094
loop_raw<float>                           719 ns        719 ns     935566
assign<float>                            1118 ns       1117 ns     576208
assign_view<float>                       1599 ns       1598 ns     384725
assign_dynamic_view<float>               6805 ns       6804 ns     103527
assign_view_noalias<float>                889 ns        889 ns     819634
assign_dynamic_view_noalias<float>       4626 ns       4625 ns     15029
```