@SylvainCorlay 2018-03-16
@ukoethe let me know if this improves iterator performances in you set up.